Finished project.

+++ Zadanie6/Services/AnimalsService.cs
@@ -13,9 +13,9 @@ public class AnimalsService : IAnimalsService
     }
 
 
-    public IEnumerable<Animal> GetAnimals()
+    public IEnumerable<Animal> GetAnimals(string orderBy)
     {
-        return _animalsRepository.GetAnimals();
+        return _animalsRepository.GetAnimals(orderBy);
     }
 
     public int CreateAnimal(Animal animal)
@@ -36,6 +36,11 @@ public class AnimalsService : IAnimalsService
 
     public int DeleteAnimal(int idAnimal)
     {
+        var existingAnimal = _animalsRepository.GetAnimalById(idAnimal);
+        if (existingAnimal == null)
+        {
+            throw new Exception("Animal with the given id does not exist.");
+        }
         return _animalsRepository.DeleteAnimal(idAnimal);
     }
